+++ scripts/bert/pretraining_utils.py
@@ -17,12 +17,17 @@
 # specific language governing permissions and limitations
 # under the License.
 # pylint:disable=redefined-outer-name,logging-format-interpolation
+import glob
+import time
 import os
 import logging
-import gluonnlp as nlp
 import mxnet as mx
+from mxnet.gluon.data import DataLoader
 
-__all__ = ['get_model']
+import gluonnlp as nlp
+from gluonnlp.data.batchify import Tuple, Stack, Pad
+
+__all__ = ['get_model', 'get_pretrain_dataset', 'get_dummy_dataloader']
 
 def get_model(ctx, model, pretrained, dataset_name, dtype, ckpt_dir=None, start_step=None):
     """Get model for pre-training."""
@@ -49,3 +54,73 @@ def get_model(ctx, model, pretrained, dataset_name, dtype, ckpt_dir=None, start_
     mlm_loss.hybridize(static_alloc=True)
 
     return model, nsp_loss, mlm_loss, vocabulary
+
+def get_pretrain_dataset(data, batch_size, num_ctxes, shuffle, use_avg_len,
+                         num_buckets, num_parts=1, part_idx=0, prefetch=True):
+    """create dataset for pretraining."""
+    num_files = len(glob.glob(os.path.expanduser(data)))
+    assert num_files >= num_parts, \
+        'Number of training files must be greater than the number of partitions'
+    split_sampler = nlp.data.SplitSampler(num_files, num_parts=num_parts, part_index=part_idx)
+    stream = nlp.data.SimpleDatasetStream(nlp.data.NumpyDataset, data, split_sampler)
+    if prefetch:
+        stream = nlp.data.PrefetchingStream(stream)
+
+    def get_dataloader(dataset):
+        """create data loader based on the dataset chunk"""
+        t0 = time.time()
+        lengths = dataset.get_field('valid_lengths')
+        logging.debug('Num samples = %d', len(lengths))
+        # A batch includes: input_id, masked_id, masked_position, masked_weight,
+        #                   next_sentence_label, segment_id, valid_length
+        batchify_fn = Tuple(Pad(), Pad(), Pad(), Pad(), Stack(), Pad(), Stack())
+        if use_avg_len:
+            # sharded data loader
+            sampler = nlp.data.FixedBucketSampler(lengths=lengths,
+                                                  # batch_size per shard
+                                                  batch_size=batch_size,
+                                                  num_buckets=num_buckets,
+                                                  shuffle=shuffle,
+                                                  use_average_length=True,
+                                                  num_shards=num_ctxes)
+            dataloader = nlp.data.ShardedDataLoader(dataset,
+                                                    batch_sampler=sampler,
+                                                    batchify_fn=batchify_fn,
+                                                    num_parts=num_ctxes)
+        else:
+            sampler = nlp.data.FixedBucketSampler(lengths,
+                                                  batch_size=batch_size * num_ctxes,
+                                                  num_buckets=num_buckets,
+                                                  ratio=0,
+                                                  shuffle=shuffle)
+            dataloader = DataLoader(dataset=dataset,
+                                    batch_sampler=sampler,
+                                    batchify_fn=batchify_fn,
+                                    num_workers=1)
+        logging.debug('Batch Sampler:\n%s', sampler.stats())
+        t1 = time.time()
+        logging.debug('Dataloader creation cost = %.2f s', t1 - t0)
+        return dataloader
+
+    stream = stream.transform(get_dataloader)
+    return stream
+
+def get_dummy_dataloader(dataloader, target_shape):
+    """Return a dummy data loader which returns a fixed data batch of target shape"""
+    data_iter = enumerate(dataloader)
+    _, data_batch = next(data_iter)
+    logging.debug('Searching target batch shape: %s', target_shape)
+    while data_batch[0].shape != target_shape:
+        logging.debug('Skip batch with shape %s', data_batch[0].shape)
+        _, data_batch = next(data_iter)
+    logging.debug('Found target dummy batch.')
+
+    class DummyIter():
+        def __init__(self, batch):
+            self._batch = batch
+
+        def __iter__(self):
+            while True:
+                yield self._batch
+
+    return DummyIter(data_batch)
