+++ scripts/bert/run_hvd.sh
@@ -4,9 +4,4 @@ python -c "import mxnet; print('successfully imported mxnet', mxnet); import glu
 DATA='/home/ec2-user/generated-book-feb-uncased-py3-128/train/part-0/part-000.npz'
 #DATA='/home/ec2-user/generated-enwiki-feb-uncased-py3-512/train/part-0/part-000.npz'
 
-horovodrun -np 8 -H localhost:8 python run_pretraining_hvd.py --batch_size 32 --accumulate 1 --max_len 128 --lr 1e-4 --data $DATA --warmup_ratio 0.01 --num_steps 1000000 --log_interval=50 --ckpt_dir './ckpt' --ckpt_interval 25000 --num_buckets 10 --dtype float16 #--verbose --profile gpu_0_hvd.json
-#mpirun -np 1 -H localhost:1 --bind-to none --map-by slot -mca pml ob1 -mca btl ^openib -mca btl_tcp_if_exclude docker0,lo -x NCCL_MIN_NRINGS=4 -x MXNET_USE_OPERATOR_TUNING=0 -x HOROVOD_HIERARCHICAL_ALLREDUCE=0 python run_pretraining_hvd.py --batch_size 32 --accumulate 1 --max_len 128 --lr 1e-4 --data $DATA --warmup_ratio 0.01 --num_steps 30000 --log_interval=50 --ckpt_dir './ckpt' --ckpt_interval 25000 --num_buckets 10 --dtype float16 #--profile gpu_0_no_hvd.json
-#python run_pretraining_hvd.py --batch_size 32 --accumulate 1 --max_len 128 --lr 1e-4 --data $DATA --warmup_ratio 0.01 --num_steps 30000 --log_interval=50 --ckpt_dir './ckpt' --ckpt_interval 25000 --num_buckets 10 --dtype float16 #--profile gpu_0_no_hvd.json
-#horovodrun -np 16 -H localhost:8,ip-172-31-3-108:8 python run_pretraining_hvd.py --batch_size 8 --accumulate 4 --max_len 512 --lr 1e-4 --data $DATA --warmup_ratio 0.01 --num_steps 1000000 --log_interval=50 --ckpt_dir './ckpt' --ckpt_interval 25000 --num_buckets 10 --dtype float16 #--verbose --profile gpu_0_hvd_large.json
-#mpirun -np 16 -H localhost:8,ip-172-31-3-108:8 --bind-to none --map-by slot -mca pml ob1 -mca btl ^openib -mca btl_tcp_if_exclude docker0,lo -x NCCL_MIN_NRINGS=16 -x MXNET_USE_OPERATOR_TUNING=0 -x HOROVOD_HIERARCHICAL_ALLREDUCE=1 -x MXNET_CPU_PRIORITY_NTHREADS=1 -x MXNET_GPU_WORKER_NTHREADS=2 -x MXNET_GPU_PRIORITY_NTHREADS=4 python run_pretraining_hvd.py --batch_size 32 --accumulate 1 --max_len 128 --lr 1e-4 --data $DATA --warmup_ratio 0.01 --num_steps 30000 --log_interval=50 --ckpt_dir './ckpt' --ckpt_interval 25000 --num_buckets 10 --dtype float16 #--profile hvd_2_node.json
-#mpirun -np 64 -H localhost:8,ip-172-31-3-108:8,ip-172-31-9-207:8,ip-172-31-10-211:8,172.31.0.114:8,172.31.13.20:8,172.31.8.99:8,172.31.1.109:8 --bind-to none --map-by slot -mca pml ob1 -mca btl ^openib -mca btl_tcp_if_exclude docker0,lo -x NCCL_MIN_NRINGS=2 -x MXNET_USE_OPERATOR_TUNING=0 -x HOROVOD_HIERARCHICAL_ALLREDUCE=1 python run_pretraining_hvd.py --batch_size 32 --accumulate 1 --max_len 128 --lr 1e-4 --data $DATA --warmup_ratio 0.01 --num_steps 30000 --log_interval=50 --ckpt_dir './ckpt' --ckpt_interval 25000 --num_buckets 10 --dtype float16
+horovodrun -np 8 -H localhost:8 python run_pretraining_hvd.py --batch_size 32 --accumulate 1 --dummy_data_len 128 --lr 1e-4 --data $DATA --warmup_ratio 0.01 --num_steps 1000000 --log_interval=50 --ckpt_dir './ckpt' --ckpt_interval 25000 --num_buckets 10 --dtype float16 #--verbose --profile gpu_0_hvd.json
